backend/Videoflix-Backend/requirements.txt
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-import-export==3.2.0
django-redis==5.3.0
django-rq==2.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
MarkupPy==1.14
moviepy==1.0.3
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
proglog==0.1.10
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
rq==1.15.1
rq-win==0.4.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
times==0.7
tqdm==4.65.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
